Move browser source map test to interactive tests.
This test tries to manually launch a browser for verification. This causes
issues when trying to launch a browser from headless mode over ssh. The
test also doesn't launch the correct file URL in windowed mode, so I'm
guessing this hasn't been check for a very long time.
